fix pre-commit exceptions
