+++ .pre-commit-config.yaml
@@ -23,7 +23,7 @@ repos:
     hooks:
       - id: flake8
         additional_dependencies: ['flake8-copyright']
-        exclude: (doc/|.github/|pyrit/prompt_converter/morse_converter.py|tests/converter/test_morse_converter.py)
+        exclude: (doc/|.github/|pyrit/prompt_converter/morse_converter.py|tests/converter/test_prompt_converter.py)
 
   - repo: local
     hooks:
