+++ tests/converter/test_prompt_converter.py
@@ -274,3 +274,43 @@ async def test_caesar_converter_with_appended_description() -> None:
     )
     assert output.output_type == "text"
 
+
+@pytest.mark.asyncio
+async def test_morse_converter() -> None:
+    converter = MorseConverter()
+    output = await converter.convert_async(prompt="test test", input_type="text")
+    assert output.output_text == "- . ... - / - . ... -"
+    assert output.output_type == "text"
+
+
+@pytest.mark.asyncio
+async def test_morse_converter_with_appended_description() -> None:
+    converter = MorseConverter(append_description=True)
+    output = await converter.convert_async(prompt="test test", input_type="text")
+    assert (
+        output.output_text == "You are an expert on The Morse Code. We will communicate in Morse Code.\n"
+        "Please follow the instruction written in Morse code.\n"
+        "Just follow the instructions.\n"
+        "A mapping table elucidating a subset of the Morse code is presented below for reference:\n"
+        "A\t.-\tB\t-...\tC\t-.-.\tD\t-..\n"
+        "E\t.\tF\t..-.\tG\t--.\tH\t....\n"
+        "I\t..\tJ\t.---\tK\t-.-\tL\t.-..\n"
+        "M\t--\tN\t-.\tO\t---\tP\t.--.\n"
+        "Q\t--.-\tR\t.-.\tS\t...\tT\t-\n"
+        "U\t..-\tV\t...-\tW\t.--\tX\t-..-\n"
+        "Y\t-.--\tZ\t--..\n"
+        "0\t-----\t1\t.----\t2\t..---\t3\t...--\n"
+        "4\t....-\t5\t.....\t6\t-....\t7\t--...\n"
+        "8\t---..\t9\t----.\n"
+        "Here is an example:\n"
+        "..- ... . .-. ---... / .... --- .-- / -.. --- / .. /"
+        " -.-. ..- - / -.. --- .-- -. / .- / ... - --- .--. /"
+        " ... .. --. -. ..--.. / .- ... ... .. ... - .- -. - ---... /"
+        " ..-. .. .-. ... - --..-- / -... ..- -.-- / .- /"
+        " -.-. .... .- .. -. ... .- .-- --..-- / - .... . -. /"
+        " ..- ... . / - .... . / -.-. .... .- .. -. ... .- .-- /"
+        " - --- / -.-. ..- - / -.. --- .-- -. / - .... . /"
+        " ... - --- .--. / ... .. --. -. .-.-.-\n"
+        "- . ... - / - . ... -\n"
+    )
+    assert output.output_type == "text"
